KS-44 #comment change file and class names
