+++ lib/screens/home/home.dart
@@ -9,6 +9,3 @@ export 'screens/pricing/pricing.dart';
 
 // Profile
 export 'screens/profile/profile.dart';
-
-//Spending Breakdown
-export 'screens/spendingBreakdown/spending.dart';
