+++ lib/screens/home/screens/dashboard/dashboard.dart
@@ -1,2 +1,5 @@
 export 'dashboard_screen.dart';
-export 'simba_screen.dart';
+export 'simba_dash.dart';
+
+// Components
+export 'components/view_transactions.dart';
